+++ just_audio/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.10.1
 
-* 
+* Fix conversion between milliBel and deciBel (@Chaphasilor).
 
 ## 0.9.36